Revert "feat(codex): reestrutura serviço Codex e atualiza importações para nova arquitetura"
This reverts commit c7e14e9c6dd8ed48468f800334c5a127c0fdda22.
